# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "srvcs-cylindervolume",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "summary": "`GET /` — service identity (srvcs service standard).",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "`POST /` — compute the volume of a cylinder by composing two float\nprimitives.",
        "description": "This service owns the *control flow* but delegates every arithmetic step to\nits dependencies, exactly as specified (`V = pi * r^2 * height`):\n\n1. ask `srvcs-pi` (a constant service, called with an empty body) for `p`;\n2. ask `srvcs-floatmultiply` for `r2 = radius * radius`;\n3. ask `srvcs-floatmultiply` for `base = p * r2` — i.e. `pi * r^2`;\n4. ask `srvcs-floatmultiply` for `result = base * height` — `pi * r^2 * h`.\n\nIt never validates `radius`/`height` itself: validation propagates from the\ndependencies, whose `422`s are forwarded verbatim. If a dependency is\nunreachable it reports itself degraded (`503`).",
        "operationId": "evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CylinderVolumeResponse"
                }
              }
            }
          },
          "422": {
            "description": "a dependency rejected the input (forwarded)"
          },
          "500": {
            "description": "a dependency returned a malformed result"
          },
          "503": {
            "description": "a dependency is unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CylinderVolumeResponse": {
        "type": "object",
        "required": [
          "radius",
          "height",
          "result"
        ],
        "properties": {
          "height": {
            "type": "object"
          },
          "radius": {
            "type": "object"
          },
          "result": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "EvalRequest": {
        "type": "object",
        "required": [
          "radius",
          "height"
        ],
        "properties": {
          "height": {
            "type": "object",
            "description": "The height of the cylinder."
          },
          "radius": {
            "type": "object",
            "description": "The radius of the cylinder's circular base."
          }
        }
      },
      "Info": {
        "type": "object",
        "required": [
          "service",
          "concern",
          "depends_on"
        ],
        "properties": {
          "concern": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "string"
          }
        }
      }
    }
  }
}